hpd v0.0.1 Hpd.Seeds

Provides functions to facilitate importing CSV data to the database.

Functions

clean_data(data)

Cleans given data by fixing malformed entries.

Examples

iex> Hpd.Seeds.clean_data %{"recommended_osVersion" => ""}
%{"recommended_osVersion" => "-"}

iex> Hpd.Seeds.clean_data %{"installDate" => ""}
%{"installDate" => "2001-01-01T01:01:01Z"}

iex> Hpd.Seeds.clean_data %{"foo" => ""}
%{"foo" => 0}

iex> Hpd.Seeds.clean_data %{"bar" => "baz"}
%{"bar" => "baz"}

clean_fields(fields)

Cleans a list of fields by trimming them and replacing all periods with underscores.

decode_file(path)

Decodes a given CSV path to a list of rows.

get_data_files()

Gets a list of the CSV data files in the /data directory.

insert_system(mapping)

map_fields(fields, data_row)

Maps a list of fields to their corresponding data elements from data_row.